requirement.txt
Django==1.11
djangorestframework==3.7.3
